feat: rebuild channel when disconnected
